+++ src/prefect/agent/fargate/__init__.py
@@ -1,0 +1,1 @@
+from prefect.agent.fargate.agent import FargateAgent