+++ src/prefect/agent/kubernetes/agent.py
@@ -40,11 +40,13 @@ class KubernetesAgent(Agent):
         from kubernetes import client, config
 
         try:
+            self.logger.debug("Loading incluster configuration")
             config.load_incluster_config()
         except config.config_exception.ConfigException as exc:
             self.logger.warning(
                 "{} Using out of cluster configuration option.".format(exc)
             )
+            self.logger.debug("Loading out of cluster configuration")
             config.load_kube_config()
 
         self.batch_client = client.BatchV1Api()
@@ -57,6 +59,9 @@ class KubernetesAgent(Agent):
             - flow_runs (list): A list of GraphQLResult flow run objects
         """
         for flow_run in flow_runs:
+            self.logger.debug(
+                "Deploying flow run {}".format(flow_run.id)  # type: ignore
+            )
 
             # Require Docker storage
             if not isinstance(StorageSchema().load(flow_run.flow.storage), Docker):
@@ -67,6 +72,9 @@ class KubernetesAgent(Agent):
 
             job_spec = self.replace_job_spec_yaml(flow_run)
 
+            self.logger.debug(
+                "Creating namespaced job {}".format(job_spec["metadata"]["name"])
+            )
             self.batch_client.create_namespaced_job(
                 namespace=os.getenv("NAMESPACE", "default"), body=job_spec
             )
@@ -104,6 +112,12 @@ class KubernetesAgent(Agent):
             StorageSchema().load(flow_run.flow.storage).name  # type: ignore
         )
 
+        self.logger.debug(
+            "Using image {} for job".format(
+                StorageSchema().load(flow_run.flow.storage).name  # type: ignore
+            )
+        )
+
         # Populate environment variables for flow run execution
         env = job["spec"]["template"]["spec"]["containers"][0]["env"]
 
@@ -124,6 +138,7 @@ class KubernetesAgent(Agent):
         token: str = None,
         api: str = None,
         namespace: str = None,
+        image_pull_secrets: str = None,
         resource_manager_enabled: bool = False,
     ) -> str:
 
@@ -143,6 +158,7 @@ class KubernetesAgent(Agent):
         agent_env[1]["value"] = api
         agent_env[2]["value"] = namespace
 
+        # Populate resource manager if requested
         if resource_manager_enabled:
             resource_manager_env = deployment["spec"]["template"]["spec"]["containers"][
                 1
@@ -153,6 +169,14 @@ class KubernetesAgent(Agent):
             resource_manager_env[3]["value"] = namespace
         else:
             del deployment["spec"]["template"]["spec"]["containers"][1]
+
+        # Populate image pull secrets if provided
+        if image_pull_secrets:
+            agent_env = deployment["spec"]["template"]["spec"]["imagePullSecrets"][0][
+                "name"
+            ] = image_pull_secrets
+        else:
+            del deployment["spec"]["template"]["spec"]["imagePullSecrets"]
 
         return yaml.safe_dump(deployment)