+++ src/prefect/cli/agent.py
@@ -5,6 +5,7 @@ from prefect.utilities.configuration import set_temporary_config
 from prefect.utilities.serialization import from_qualified_name
 
 _agents = {
+    "fargate": "prefect.agent.fargate.FargateAgent",
     "local": "prefect.agent.local.LocalAgent",
     "kubernetes": "prefect.agent.kubernetes.KubernetesAgent",
     "nomad": "prefect.agent.nomad.NomadAgent",
@@ -46,26 +47,37 @@ def agent():
 @click.option(
     "--token", "-t", required=False, help="A Prefect Cloud API token.", hidden=True
 )
+@click.option(
+    "--verbose", "-v", is_flag=True, help="Enable verbose agent logs.", hidden=True
+)
 @click.option("--no-pull", is_flag=True, help="Pull images flag.", hidden=True)
 @click.option("--base-url", "-b", help="Docker daemon base URL.", hidden=True)
-def start(name, token, no_pull, base_url):
+def start(name, token, verbose, no_pull, base_url):
     """
     Start an agent.
 
     \b
     Arguments:
-        name            TEXT    The name of an agent to start (e.g. `local`, `kubernetes`, `nomad`)
+        name            TEXT    The name of an agent to start (e.g. `local`, `kubernetes`, `fargate`, `nomad`)
                                 Defaults to `local`
 
     \b
     Options:
         --token, -t     TEXT    A Prefect Cloud API token with RUNNER scope
+        --verbose, -v           Enable verbose agent DEBUG logs
+                                Defaults to INFO level logging
+
+    \b
+    Local Agent Options:
         --base-url, -b  TEXT    A Docker daemon host URL for a LocalAgent
         --no-pull               Pull images for a LocalAgent
                                 Defaults to pulling if not provided
     """
     with set_temporary_config(
-        {"cloud.agent.auth_token": token or config.cloud.agent.auth_token}
+        {
+            "cloud.agent.auth_token": token or config.cloud.agent.auth_token,
+            "cloud.agent.level": "INFO" if not verbose else "DEBUG",
+        }
     ):
         retrieved_agent = _agents.get(name, None)
 
@@ -93,24 +105,32 @@ def start(name, token, no_pull, base_url):
     hidden=True,
 )
 @click.option(
+    "--image-pull-secrets",
+    "-i",
+    required=False,
+    help="Name of image pull secrets to use for workloads.",
+    hidden=True,
+)
+@click.option(
     "--resource-manager", is_flag=True, help="Enable resource manager.", hidden=True
 )
-def install(name, token, api, namespace, resource_manager):
+def install(name, token, api, namespace, image_pull_secrets, resource_manager):
     """
     Install an agent. Outputs configuration text which can be used to install on various
     platforms.
 
     \b
     Arguments:
-        name                TEXT    The name of an agent to start (e.g. `kubernetes`)
-                                    Defaults to `kubernetes`
+        name                        TEXT    The name of an agent to start (e.g. `kubernetes`)
+                                            Defaults to `kubernetes`
 
     \b
     Options:
-        --token, -t         TEXT    A Prefect Cloud API token
-        --api, -a           TEXT    A Prefect Cloud API URL
-        --namespace, -n     TEXT    Agent namespace to launch workloads
-        --resource-manager          Enable resource manager on install
+        --token, -t                 TEXT    A Prefect Cloud API token
+        --api, -a                   TEXT    A Prefect Cloud API URL
+        --namespace, -n             TEXT    Agent namespace to launch workloads
+        --image-pull-secrets, -i    TEXT    Name of image pull secrets to use for workloads
+        --resource-manager                  Enable resource manager on install
     """
 
     supported_agents = {"kubernetes": "prefect.agent.kubernetes.KubernetesAgent"}
@@ -125,6 +145,7 @@ def install(name, token, api, namespace, resource_manager):
         token=token,
         api=api,
         namespace=namespace,
+        image_pull_secrets=image_pull_secrets,
         resource_manager_enabled=resource_manager,
     )
     click.echo(deployment)